pyproject.toml
[tool.poetry]
name = "glasses"
version = "0.1.0"
description = "A simple Django web app for viewing and purchasing frames and lenses."
authors = ["aka-radix <h.radiahmad@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.4"
djangorestframework = "^3.15.1"
django-environ = "^0.11.2"
dj-rest-auth = "^6.0.0"
django-allauth = "^0.62.1"
djangorestframework-simplejwt = "^5.3.1"
django-structlog = "^8.0.0"
drf-spectacular = "^0.27.2"
psycopg2-binary = "^2.9.9"


[tool.poetry.group.dev.dependencies]
ruff = "^0.4.2"
mypy = "^1.10.0"
isort = "^5.13.2"
autoflake = "^2.3.1"
flake8 = "^7.0.0"
djlint = "^1.34.1"
django-silk = "^5.1.0"
django-extensions = "^3.2.3"
ipython = "^8.24.0"
nplusone = "^1.0.0"
pre-commit = "^3.7.0"

[tool.mypy]
python_version = "^3.12"
check_untyped_defs = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
plugins = [
    "mypy_django_plugin.main",
    "mypy_drf_plugin.main",
]

[[tool.mypy.overrides]]
module = "*.migrations.*"
ignore_errors = true

[tool.django-stubs]
django_settings_module = "config.settings"

[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "*/migrations/*.py",
    "staticfiles/*"
]
line-length = 79
indent-width = 4
target-version = "py312"

[tool.ruff.lint]
select = [
    "I",
    "F",
    "E",
    "W",
    "C90",
    "I",
    "N",
    "UP",
    "YTT",
    "ASYNC",
    "S",
    "BLE",
    "FBT",
    "B",
    "A",
    "COM",
    "C4",
    "DTZ",
    "T10",
    "DJ",
    "EM",
    "EXE",
    "FA",
    'ISC',
    "ICN",
    "G",
    'INP',
    'PIE',
    "T20",
    'PYI',
    'PT',
    "Q",
    "RSE",
    "RET",
    "SLF",
    "SLOT",
    "SIM",
    "TID",
    "TCH",
    "INT",
    "ARG",
    "PTH",
    "ERA",
    "PD",
    "PGH",
    "PL",
    "TRY",
    "FLY",
    "PERF",
    "RUF"
]

ignore = [
    "S101",
    "RUF012",
    "SIM102",
    "TRY003",
    "EM101",
    "S311",
    "UP007",
    "E731",
    "PGH003",
    "SLF001",
    "TRY301",
    "ARG001",
    "ARG002",
    "FBT002",
]

fixable = ["ALL"]
unfixable = []

dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.isort]
force-single-line = false

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
